# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Informasi Manajemen Inventaris Sekolah API",
        "description": "REST API untuk Sistem Informasi Manajemen Inventaris Sekolah. Mendukung modul Auth, Pengguna, Peran & Akses, Master Sekolah, Master Barang, Manajemen Aset, Transaksi Peminjaman & Permintaan, Pemeliharaan & Akhir Siklus Aset, serta Manajemen Database.",
        "contact": {
            "name": "Admin Sekolah",
            "email": "admin@sekolah.sch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/ruang": {
            "get": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Daftar semua ruang",
                "operationId": "indexRuang",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuangListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Tambah ruang baru",
                "operationId": "storeRuang",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRuangRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuangSingleResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ruang/{id}": {
            "get": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Detail ruang",
                "operationId": "showRuang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuangSingleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Update ruang",
                "operationId": "updateRuang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRuangRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuangSingleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ruang"
                ],
                "summary": "Hapus ruang",
                "operationId": "destroyRuang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuangDeleteResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LokasiResource": {
                "description": "Representasi data lokasi",
                "properties": {
                    "id_lokasi": {
                        "description": "============================================================\n SCHEMA DEFINITIONS – Lokasi Module\n============================================================",
                        "type": "integer",
                        "example": 1
                    },
                    "nama_lokasi": {
                        "type": "string",
                        "example": "Gedung A"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Pendidikan No.1",
                        "nullable": true
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Utama",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "RuangResource": {
                "properties": {
                    "id_ruang": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_ruang": {
                        "type": "string",
                        "example": "Lab Komputer 1"
                    },
                    "id_lokasi": {
                        "type": "integer",
                        "example": 1
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Ruangan ber-AC",
                        "nullable": true
                    },
                    "lokasi": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LokasiResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StoreRuangRequest": {
                "required": [
                    "nama_ruang",
                    "id_lokasi"
                ],
                "properties": {
                    "nama_ruang": {
                        "type": "string",
                        "example": "Lab Komputer 1",
                        "maxLength": 100
                    },
                    "id_lokasi": {
                        "type": "integer",
                        "example": 1
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Ruangan ber-AC",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateRuangRequest": {
                "properties": {
                    "nama_ruang": {
                        "type": "string",
                        "example": "Lab Komputer 2",
                        "maxLength": 100
                    },
                    "id_lokasi": {
                        "type": "integer",
                        "example": 1
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Ruangan ber-AC",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RuangListResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Daftar ruang berhasil diambil."
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RuangResource"
                        }
                    }
                },
                "type": "object"
            },
            "RuangSingleResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Detail ruang berhasil diambil."
                    },
                    "data": {
                        "$ref": "#/components/schemas/RuangResource"
                    }
                },
                "type": "object"
            },
            "RuangDeleteResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Ruang berhasil dihapus."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}